@@ -1,14 +1,14 @@
 {
-  "name": "@project-sunbird/sunbird-pdf-player",
-  "version": "0.1.0",
+  "name": "@project-sunbird/sunbird-pdf-player-v8",
+  "version": "0.0.1",
   "peerDependencies": {
-    "@angular/common": "^7.2.0",
-    "@angular/core": "^7.2.0",
+    "@angular/common": "^8.0.0",
+    "@angular/core": "^8.0.0",
     "@project-sunbird/client-services": "^3.4.8",
     "@project-sunbird/sb-styles": "0.0.7"
   },
   "dependencies": {
-    "@project-sunbird/sunbird-player-sdk": "0.0.13"
+    "@project-sunbird/sunbird-player-sdk-v8": "0.0.1"
   },
   "devDependencies": {
     "@project-sunbird/client-services": "^3.4.8",
