@@ -10,7 +10,7 @@ import { ViewerService } from '../services/viewer.service';
 export class PdfViewerComponent implements AfterViewInit, OnDestroy {
 
   public src = './../assets/pdfjs/web/viewer.html?file=';
-  @ViewChild('iframe') iframeRef: ElementRef;
+  @ViewChild('iframe', {static: true}) iframeRef: ElementRef;
   @Input() pdfURL: string;
   @Input() actions = new EventEmitter<any>();
   private unListenLoadEvent: () => void;
