@@ -1,7 +1,7 @@
 const fs = require('fs-extra');
 (async () => {
     try {
-      var source = "projects/sunbird-pdf-player/node_modules/@project-sunbird/sunbird-player-sdk/lib/assets";
+      var source = "projects/sunbird-pdf-player/node_modules/@project-sunbird/sunbird-player-sdk-v8/lib/assets";
         const dest = "dist/sunbird-pdf-player/lib/assets/";
         var libsource = "projects/sunbird-pdf-player/src/lib/assets";
         const isAssetsExists = await fs.pathExists(dest)
